cleaned up code and used smaller data types

+++ Assets/Game/Scripts/EventScripts/CaptureTheFlag.cs
@@ -9,7 +9,7 @@ public class CaptureTheFlag : GameEvent
     public Flag flag;
     public GameObject flagSpawnpoint;
     public GameObject[] endPoints;
-    public int pointsForCapture;
+    public byte pointsForCapture;
     public float flagResetTime;
 
     GameObject carrier;
